Extract strategy endpoints into route and service layers
